Required Assets
These were a part of the original files that I had cloned from somewhere. If someone can help me with finding who did this, I want to be able to properly give them credit. I'm pretty sure it was from Huggingface, and from the looks of it, it's taken a lot of the "src" folder's files from the original ComfyBox, and pulled a javascript from it.

I tried to run without this Folder and had the same issues others were mentioning (ie blank screen)